requirements.txt
# This repo uses Python 3.9.16
# Create a requirements.txt file with the following command: 
# pip freeze > requirements.txt
# You'll notice that this also includes all dependencies of the packages once installed.
# Try to clean up a bit.

lightning==2.1.3
lightning-utilities==0.10.0
matplotlib==3.8.2
mpmath==1.3.0
numpy==1.26.3
pandas==2.1.4
patsy==0.5.6
plotly==5.18.0
pytorch-lightning==2.1.3
PyYAML==6.0.1
SciencePlots==2.1.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
statsmodels==0.14.1
threadpoolctl==3.2.0
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
